# backend/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Project Social Network API",
      "description": "Project Social Network API - by Bootcamp FullStack SysMap",
      "contact": {
        "email": "alexandre.xoquito@gmail.com"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:4000/v1"
      }
    ],
...
    "paths": {
      "/security/register": {
        "post": {
          "tags": ["security"],
          "summary": "Creates a new user",
          "description": "Creates a new user",
          "operationId": "registerUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "CREATED",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RegisterResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorArrayResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/security/login": {
        "post": {
          "tags": ["security"],
          "summary": "User's login",
          "description": "Login a user",
          "operationId": "logInUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LoginResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "404": {
              "description": "NOT FOUND",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/posts": {
        "get": {
          "tags": ["posts"],
          "summary": "Get all existing posts",
          "description": "Get all existing posts",
          "operationId": "getPosts",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "post": {
          "tags": ["posts"],
          "summary": "Creates a new post",
          "description": "Creates a new post",
          "operationId": "createNewPost",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "CREATED",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorArrayResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/posts/{id}": {
        "get": {
          "tags": ["posts"],
          "summary": "Find posts by id",
          "description": "Find posts by id",
          "operationId": "findPostById",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to get"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "put": {
          "tags": ["posts"],
          "summary": "Update a post",
          "description": "This can only be done by the logged in user.",
          "operationId": "updatePost",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to update"
            }
          ],
          "requestBody": {
            "description": "Update an existent post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRequest"
                }
              }
            }
          },
          "responses": {
            "203": {
              "description": "UNDOCUMENTED",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "delete": {
          "tags": ["posts"],
          "summary": "Deletes a post",
          "description": "delete a post",
          "operationId": "deletePost",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to delete"
            }
          ],
          "responses": {
            "204": {
              "description": "NO CONTENT"
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/posts/{id}/like": {
        "post": {
          "tags": ["posts"],
          "summary": "Likes the post",
          "description": "Likes the post, liking the same post twice will result in a bad request",
          "operationId": "likesPost",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to like"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponseLikeUnlike"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST"
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/posts/{id}/unlike": {
        "post": {
          "tags": ["posts"],
          "summary": "Unlikes the post",
          "description": "Unlikes the post, unliking the same post twice will result in a bad request",
          "operationId": "unlikesPost",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to like"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PostResponseLikeUnlike"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST"
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/posts/{postId}/comments": {
        "get": {
          "tags": ["comments"],
          "summary": "Get all comments from a post",
          "description": "Get all comments in post",
          "operationId": "getPosts",
          "parameters": [
            {
              "in": "path",
              "name": "postId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to get comments from"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CommentResponse"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ErrorArrayResponse"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "post": {
          "tags": ["comments"],
          "summary": "Create a new comment",
          "description": "Create a new comment",
          "operationId": "createNewComment",
          "parameters": [
            {
              "in": "path",
              "name": "postId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the post to get comments from"
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "CREATED",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CommentPostResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ErrorArrayResponse"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
...
      "/profiles": {
        "get": {
          "tags": ["profile"],
          "summary": "Find all profiles",
          "description": "Find all profiles",
          "operationId": "findProfiles",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ProfileResponse"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/profiles/search?q={q}": {
        "get": {
          "tags": ["profile"],
          "summary": "Search for a profile by its name",
          "description": "Search for a profile by its name",
          "operationId": "findProfileByName",
          "parameters": [
            {
              "in": "query",
              "name": "q",
              "schema": {
                "type": "string",
                "example": "Kelly Castelo"
              },
              "required": true,
              "description": "profile's name"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/profile/:id/follow": {
        "post": {
          "tags": ["profile"],
          "summary": "Follow a profile",
          "description": "Follow a profile, returns 400 if you're already following this profile",
          "operationId": "followProfile",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the profile to follow"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST"
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
      "/profile/:id/unfollow": {
        "post": {
          "tags": ["profile"],
          "summary": "Unfollow a profile",
          "description": "Unfollow a profile, returns 400 if you're not following this profile",
          "operationId": "unfollowProfile",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "uuid of the profile to unfollow"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  }
                }
              }
            },
            "400": {
              "description": "BAD REQUEST"
            },
            "401": {
              "description": "UNAUTHORIZED"
            },
            "404": {
              "description": "NOT FOUND"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },      
...
    },
    "components": {
      "schemas": {
        "LoginRequest": {
          "type": "object",
          "properties": {
            "user": {
              "type": "string",
              "minLength": 2,
              "example": "lizzera"
            },
            "password": {
              "type": "string",
              "minLength": 2,
              "example": "q1w2e3r4"
            }
          }
        },
        "LoginResponse": {
          "type": "object",
          "properties": {
            "token": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiJ9.eyJfaWQiOiI2Mzg3OWM3ODJmN2RhZjdmYmUwYTlmMjkiLCJ1c2VyIjoianVuaW1hbCIsIl9fdiI6MCwicHJvZmlsZSI6eyJfaWQiOiI2Mzg3OWM3ODJmN2RhZjdmYmUwYTlmMmIiLCJuYW1lIjoiSnVuaW9yIiwidXNlciI6IjYzODc5Yzc4MmY3ZGFmN2ZiZTBhOWYyOSIsImZvbGxvd2luZyI6W10sImZvbGxvd2VycyI6W10sIl9fdiI6MH19.aTXFGQ4lHZCah6_MSpiHOZH1DmyDXmlCmwZHTMuyE6k"
            }
          }
        },
        "RegisterRequest": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "minLength": 2,
              "example": "Alexandre Siqueira"
            },
            "user": {
              "type": "string",
              "minLength": 2,
              "example": "lizzera"
            },
            "password": {
              "type": "string",
              "minLength": 2,
              "example": "q1w2e3r4"
            }
          }
        },
        "RegisterResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879d1e2f7daf7fbe0a9f32"
            },
            "user": {
              "type": "string",
              "example": "lizzera"
            }
          }
        },
        "ErrorArrayResponse": {
          "type": "object",
          "properties": {
            "path": {
              "type": "string",
              "example": "/v1/path/action"
            },
            "status": {
              "type": "integer",
              "example": "400"
            },
            "errors": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "Path `user` is required"
              }
            }
          }
        },
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "path": {
              "type": "string",
              "example": "/v1/path/action"
            },
            "status": {
              "type": "integer"
            },
            "errors": {
              "type": "string"
            }
          }
        },
        "PostResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879d1e2f7daf7fbe0a9f32"
            },
            "title": {
              "type": "string",
              "example": "World Cup"
            },
            "description": {
              "type": "string",
              "example": "Brazilians best football"
            },
            "profile": {
              "$ref": "#/components/schemas/ProfileResponse"
            },
            "comments": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "6387a0392f7daf7fbe0a9f5a"
              }
            },
            "likes": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "63879c782f7daf7fbe0a9f2b"
              }
            }
          }
        },
        "PostResponseLikeUnlike": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879d1e2f7daf7fbe0a9f32"
            },
            "title": {
              "type": "string",
              "example": "What's the Champion"
            },
            "description": {
              "type": "string",
              "example": "Use the imagination"
            },
            "profile": {
              "type": "string",
              "example": "63879c782f7daf7fbe0a9f2b"
            },
            "comments": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "6387a0392f7daf7fbe0a9f5a"
              }
            },
            "likes": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "637bb9d4da4e28ea8606b7c4"
              }
            }
          }
        },
        "PostRequest": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "example": "World Cup"
            },
            "description": {
              "type": "string",
              "example": "Use the imagination"
            }
          },
          "required": ["title", "description"]
        },
        "ProfileResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879c782f7daf7fbe0a9f2b"
            },
            "name": {
              "type": "string",
              "example": "Alexandre Siqueira"
            },
            "user": {
              "type": "string",
              "example": "Lizzera"
            },
            "following": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "63879c782f7daf7fbe0a9f2b"
              }
            },
            "followers": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "637bb9d4da4e28ea8606b7c8"
              }
            }
          }
        },
...
        "CommentRequest": {
          "type": "object",
          "properties": {
            "description": {
              "type": "string",
              "example": "Nice Post!"
            }
          },
          "required": ["description"]
        },
        "CommentResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879d1e2f7daf7fbe0a9f32"
            },
            "description": {
              "type": "string",
              "example": "Use the imagination"
            },
            "profile": {
              "$ref": "#/components/schemas/ProfileResponse"
            },
            "comments": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "6387a0392f7daf7fbe0a9f5a"
              }
            },
            "likes": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "63879c782f7daf7fbe0a9f2b"
              }
            }
          }
        },
        "CommentPostResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "637bbc3a69c083c2b0003097"
            },
            "description": {
              "type": "string",
              "example": "Very nice post"
            },
            "profile": {
              "type": "string",
              "example": "63879c782f7daf7fbe0a9f2b"
            },
            "comments": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "6387a0392f7daf7fbe0a9f5a"
              }
            },
            "likes": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "63879c782f7daf7fbe0a9f2b"
              }
            }
          }
        }
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    }
  }